/trunk/core  * separation of effects (classIndex dependent) and skill reuses (classIndex independent)
git-svn-id: http://svn.l2jfree.com/l2jfree@7706 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/model/actor/L2Character.java
@@ -6334,7 +6334,7 @@ public abstract class L2Character extends L2Object
 	 * @param delay in milliseconds
 	 * @return modified
 	 */
-	public boolean disableSkill(int skillId, long delay)
+	public boolean disableSkill(int skillId, int delay)
 	{
 		if (delay < 100)
 			return false;
